fix: address mobile task list review feedback
