+++ apps/web/components/kanban/kanban-header-mobile.tsx
@@ -6,7 +6,8 @@ import Link from "@/components/routing/app-link";
 import { PageTopbar } from "@/components/page-topbar";
 import { TopbarMetrics } from "@/components/system-metrics/topbar-metrics";
 import { MainTopBarPluginActions } from "./main-top-bar-plugin-actions";
-import { MobileMenuSheet, type TasksListDisplayOptions } from "./mobile-menu-sheet";
+import { MobileMenuSheet } from "./mobile-menu-sheet";
+import type { TasksListDisplayOptions } from "./mobile-menu-task-list-options";
 import { useAppStore } from "@/components/state-provider";
 import { useQuickChatLauncher } from "@/hooks/use-quick-chat-launcher";
 import { workspaceHomeHref } from "@/components/app-sidebar/app-sidebar-workspace-navigation";
